hides cartridge data from NES components [clang]
COMMENTS:
Makes the cartridge data (effectively) private.

This has been achieved by moving type definition of the cartridge
data type to the implementation file `cartridge.c'. Since it is
not present in the header file `cartridge.h' there is no means
for the other NES components to ``know'' the underlying type of the
private data to be able to access it.

The code displays the expected output on the console.

Valgrind reports no memory issues.