updates Cartridge::load_CHR_ROM() [clang]
COMMENTS:
If there's no VROM, then the method nullifies the VROM container
`m_CHR_ROM' and sets the VROM size `num_vbanks' to zero. That is
all that was really modified in this commit (even if *git* shows
otherwise).

This is necessary for mappers that use the VROM, for what they do
(in short) depends on the VROM size.

The emulator executes as expected.

Valgrind reports no memory issues.