adds info for getting mapper number [clang]
COMMENTS:
The added URLs explain in detail the contents of the ROM header.
I used that to make the determination of the mapper number a more
readable (hopefully).

The mapper number is stored between byte 6 and 7 (counting from 0).
The last four bits of byte 6 contain the lower nybble and the last
four bits of byte 7 contain the upper nybble of the mapper number.

A nybble is half a byte (or four bits).